test: move specs into lib-based directory structure
