+++ spec/lib/coelacanth/client_spec.rb
