+++ spec/lib/coelacanth/configure_spec.rb
