+++ spec/lib/coelacanth/dom_spec.rb
