+++ spec/lib/coelacanth/redirect.rb
