+++ spec/lib/coelacanth/validator_spec.rb
