+++ spec/lib/coelacanth_spec.rb
