+++ spec/lib/shared_examples.rb
